Handle impossible paths in the network
So, the A2 highway only had the North-South ways,
this was causing some path to be impossible in the
network. This took a whole bunch of debbuging to
figure out, now every stop is connected by at least
one projection, which means the real network lyes
somewhere in there. I also made some restructuring.